Results: improve share and new report modal styles
