+++ client/src/components/ui/alert-dialog.tsx
@@ -66,7 +66,7 @@ const AlertDialogTitle = React.forwardRef<
 >(({ className, ...props }, ref) => (
   <AlertDialogPrimitive.Title
     ref={ref}
-    className={cn("text-lg font-semibold", className)}
+    className={cn("text-lg font-bold text-foreground", className)}
     {...props}
   />
 ));
@@ -102,6 +102,7 @@ const AlertDialogCancel = React.forwardRef<
     {...props}
   />
 ));
+
 AlertDialogCancel.displayName = AlertDialogPrimitive.Cancel.displayName;
 
 export {
