+++ client/src/components/ui/dialog.tsx
@@ -58,7 +58,7 @@ const DialogClose = React.forwardRef<
     ref={ref}
     className="absolute right-4 top-4 rounded-sm opacity-70 ring-offset-background transition-opacity hover:opacity-100 focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2 disabled:pointer-events-none data-[state=open]:bg-accent data-[state=open]:text-muted-foreground"
   >
-    <LuX className="h-4 w-4" />
+    <LuX className="h-6 w-6" />
     <span className="sr-only">Close</span>
   </DialogPrimitive.Close>
 ));
