+++ client/src/i18n/translations/en.json
@@ -151,8 +151,8 @@
   "report-results-sidebar-ai-summaries-checkbox-active-indicators-info": "Include only active indicators in the generated summaries. Changes to the indicators will require generating new summaries.",
   "report-results-sidebar-ai-summaries-generating": "Generating AI summaries",
   "report-results-sidebar-ai-summaries-generating-info": "AI generated summaries can be inaccurate. We encourage you to verify the content carefully",
-  "report-results-copy-title": "Share report",
-  "report-results-copy-description": "Copy and paste link to share",
+  "report-results-copy-title": "Share & Save",
+  "report-results-copy-description": "This link lets you share the report and saves its current status. Copy it to revisit or share with others.",
   "indicators-custom-municipalities-selected-area-intersects": "The selected area intersects",
   "h3-grid-column-name": "H3 Grid Column Name",
   "other-resources-resource-info": "Resource Info",
