+++ client/src/i18n/translations/es.json
@@ -151,8 +151,8 @@
   "report-results-sidebar-ai-summaries-checkbox-active-indicators-info": "Incluir sólo los indicadores activos en los resúmenes generados. Los cambios en los indicadores requerirán la generación de nuevos resúmenes.",
   "report-results-sidebar-ai-summaries-generating": "Generación de resúmenes de IA",
   "report-results-sidebar-ai-summaries-generating-info": "Los resúmenes generados por IA pueden ser inexactos. Le recomendamos que compruebe detenidamente el contenido",
-  "report-results-copy-title": "Compartir reporte",
-  "report-results-copy-description": "Copiar y pegar enlace para compartir",
+  "report-results-copy-title": "Compartir y salvar reporte",
+  "report-results-copy-description": "Este enlace te permite compartir el reporte y guardar su estado actual. Cópialo para volver a visitarlo o compartirlo con otros.",
   "indicators-custom-municipalities-selected-area-intersects": "El área seleccionada interseca",
   "h3-grid-column-name": "Nombre de columna del Grid H3",
   "other-resources-resource-info": "Información sobre recursos",
