+++ client/src/i18n/translations/pt.json
@@ -151,8 +151,8 @@
   "report-results-sidebar-ai-summaries-checkbox-active-indicators-info": "Incluir apenas os indicadores ativos nos resumos gerados. Alterações nos indicadores exigirão a geração de novos resumos.",
   "report-results-sidebar-ai-summaries-generating": "Geração de resumos de IA",
   "report-results-sidebar-ai-summaries-generating-info": "Os resumos gerados por IA podem ser imprecisos. Recomendamos que verifique cuidadosamente o conteúdo",
-  "report-results-copy-title": "Compartilhar relatório",
-  "report-results-copy-description": "Copiar e colar link para compartilhar",
+  "report-results-copy-title": "Compartilhar e salvar relatório",
+  "report-results-copy-description": "Este link permite compartilhar o relatório e salvar seu estado atual. Copie-o para revisitá-lo ou compartilhá-lo com outros.",
   "indicators-custom-municipalities-selected-area-intersects": "A área selecionada intersecta",
   "h3-grid-column-name": "Nome da coluna do Grid H3",
   "other-resources-resource-info": "Informações sobre os recursos",
